+++ CHANGELOG.md
@@ -26,6 +26,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Removed Input component (external lib is no longer being maintained)
 - Removed Modal component (replaced by Lightbox)
 - Removed unused log util
+- Removed default android back handler disable in Lightbox component
 
 ### Changed